Move remaining CSS
